requirements.txt
#altair==4.2.0
#anaconda-navigator==2.1.1
#appdirs==1.4.4
#argh==0.26.2
#atomicwrites==1.4.0
#backports.weakref==1.0.post1
#bkcharts==0.2
#black==19.10b0
#blinker==1.4
#boto==2.49.0
#Brotli==1.0.9
#brotlipy==0.7.0
#cachetools==5.1.0
#certifi==2021.10.8
#click==8.0.3
#clyent==1.2.2
#comtypes==1.1.10
#conda==4.10.3
#conda-build==3.21.6
#conda-verify==3.4.2
#cycler==0.10.0
#cytoolz==0.11.0
#daal4py==2021.3.0
#dash==2.3.1
#dash-bootstrap-components==1.1.0
#dash-colorscales==0.0.4
#dash-core-components==2.0.0
#dash-daq==0.5.0
#dash-html-components==2.0.0
#dash-table==5.0.0
#dask==2021.10.0
#data==0.4
docx==0.2.4
docx2txt==0.8
dtale==2.2.0
#entrypoints==0.3
#et-xmlfile==1.1.0
#Flask-Compress==1.11
#flask-ngrok==0.0.25
#fonttools==4.25.0
fpdf==1.7.2
#funcsigs==1.0.2
#gitdb==4.0.9
#GitPython==3.1.27
#inflection==0.5.1
#kaleido==0.2.1
#latex==0.7.0
#llvmlite==0.37.0
#locket==0.2.1
#lz4==4.0.0
#mccabe==0.6.1
#missingno==0.4.2
#mkl-fft==1.3.1
#mkl-service==2.4.0
#mpmath==1.2.1
#munkres==1.1.4
#mypy-extensions==0.4.3
#navigator-updater==0.2.1
#nltk==3.6.5
#pathspec==0.7.0
#patsy==0.5.2
#pep8==1.7.1
#Pillow==8.4.0
#pkginfo==1.7.1
#plotly==5.7.0
#ply==3.11
#ppscore==1.2.0
#protobuf==3.20.1
#pyarrow==8.0.0
#pycosat==0.6.3
#pycurl==7.44.1
#pydeck==0.7.1
#pyls-spyder==0.4.0
#Pympler==1.0.1
#pyodbc===4.0.0-unsupported
#pyreadline==2.1
#pytest==6.2.4
#python-lsp-jsonrpc==1.0.0
#python-lsp-server==1.2.4
#pytz==2021.3
#pytz-deprecation-shim==0.1.0.post0
#pywin32==228
#PyYAML==6.0
#scikit-image==0.18.3
#semver==2.13.0
#shutilwhich==1.1.0
#simplegeneric==0.8.1
#sip==4.19.13
#smmap==5.0.0
#Sphinx==4.2.0
#squarify==0.4.3
#statsmodels==0.12.2
streamlit==1.9.0
#strsimpy==0.2.1
#tables==3.6.1
#TBB==0.2
#tempdir==0.7.1
#tenacity==8.0.1
#terminado==0.9.4
#tzdata==2022.1
#tzlocal==4.2
unicodecsv==0.14.1
#validators==0.19.0
#webencodings==0.5.1
#win-unicode-console==0.5
#wincertstore==0.2
#xarray==2022.3.0
#xlwings==0.24.9
#xlwt==1.3.0
#zict==2.0.0
#zope.event==4.5.0
altair==4.1.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-generator==1.10
attrs==21.4.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bleach==4.1.0
blinker==1.4
boto3==1.23.10
cachetools==4.2.4
certifi==2020.6.20
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cloudpickle==2.0.0
colorama==0.4.4
dataclasses==0.6
decorator==5.1.1
defusedxml==0.7.1
dm-tree==0.1.6
entrypoints==0.4
flexpolyline==0.1.0
folium==0.12.1.post1
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.18
h5py==3.1.0
idna==3.3
importlib-metadata==4.8.3
importlib-resources==5.4.0
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.17.2
Jinja2==3.0.3
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.0
Markdown==3.3.6
MarkupSafe==2.0.1
mistune==0.8.4
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.5
notebook==6.4.10
numpy==1.19.5
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parso==0.7.1
pickleshare==0.7.5
Pillow==8.4.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.19.4
pyarrow==6.0.1
pycparser==2.21
pydeck==0.6.2
Pygments==2.11.2
Pympler==1.0.1
pyparsing==3.0.8
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pyzmq==22.3.0
requests==2.27.1
semver==2.13.0
Send2Trash==1.8.0
six
smmap==5.0.0
streamlit-folium==0.4.0
tensorboard==2.6.0
tensorflow==2.6.2
tensorflow-estimator==2.6.0
tensorflow-probability==0.16.0
terminado==0.12.1
testpath==0.6.0
toml==0.10.2
toolz==0.11.2
tornado==6.1
traitlets==4.3.3
typing_extensions
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
validators==0.18.2
watchdog==2.1.7
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.6.0
wincertstore==0.2
zipp==3.6.0


